Add support for embeds and custom form field
This commit now has embeds, so you can edit embeds as well. Associations
are pretty similar to embeds, however the problem is associations can
have other associations, so how do we handle that? Only allow for one
level of associations?

Custom form field can now take a function to render input, but I'm not
sure if this is the best approach.

There's also a WIP to try to allow users to use LiveComponent, which can
encapsulate more sophisticated logic (such as a typeahead/search
suggestion dropdown kind of UI)