Fix inaccurate default widget data
The default widget shows the total entries, but then it was accidentally
relying on `resources`, which means that when a there's a filter and the
filter changes the `resources`, the total count changes to the filtered
count.

This commit fixes it by saving the total entries on mount and use only
that.