Move Single out of Index, abolish modal form.
This is a big one, we removed the popup modal for FormComponent, and
instead we require users to implement two module now: `Index` and
`Single`.

Part of the reason is because the modal has limited real estate, and
with the work to show `has_many` relationships as a table in the bottom,
we need more real estate.