Add new filtering UI
I think the end result looks pretty good, though I'm not really
satisfied with the code quality. I think there are a lot of room for
improvement here..

For example, I think there's a lot of duplication. Rendering a field
now is spread across multiple parts, we need to repeat the type logic
in `FilterComponent` and also in `Field`, can we tidy that up?

Another thing is, right now the FilterComponent "updates" the params
by navigating it to a new page - should we do that, or should we have
FilterComponent `send_update` to the parent, and let the parent do the
navigation?